+++ src/android/io/sqlc/SQLiteAndroidDatabase.java
@@ -70,6 +70,14 @@ class SQLiteAndroidDatabase
      * @param dbfile   The database File specification
      */
     void open(File dbfile) throws Exception {
+        if (!isPostHoneycomb) {
+            Log.v("SQLiteAndroidDatabase.open",
+                "INTERNAL PLUGIN ERROR: deprecated android.os.Build.VERSION not supported: " +
+                android.os.Build.VERSION.SDK_INT);
+            throw new RuntimeException(
+                "INTERNAL PLUGIN ERROR: deprecated android.os.Build.VERSION not supported: " +
+                android.os.Build.VERSION.SDK_INT);
+        }
         dbFile = dbfile; // for possible bug workaround
         mydb = SQLiteDatabase.openOrCreateDatabase(dbfile, null);
     }
@@ -143,7 +151,7 @@ class SQLiteAndroidDatabase
                 //Log.v("executeSqlBatch", "query type: " + queryType);
 
                 if (queryType == QueryType.update || queryType == queryType.delete) {
-                    if (isPostHoneycomb) {
+                    // if (isPostHoneycomb) {
                         SQLiteStatement myStatement = mydb.compileStatement(query);
 
                         if (json_params != null) {
@@ -175,7 +183,10 @@ class SQLiteAndroidDatabase
                             // Assuming SDK_INT was lying & method not found:
                             // do nothing here & try again with raw query.
                             ex.printStackTrace();
-                            Log.v("executeSqlBatch", "SQLiteStatement.executeUpdateDelete(): runtime error (fallback to old API): " + errorMessage);
+                            // Log.v("executeSqlBatch", "SQLiteStatement.executeUpdateDelete(): runtime error (fallback to old API): " + errorMessage);
+                            Log.v("SQLiteAndroidDatabase.executeSqlBatchStatement",
+                                "INTERNAL PLUGIN ERROR: could not do myStatement.executeUpdateDelete(): " + ex.getMessage());
+                            throw(ex);
                         }
 
                         // "finally" cleanup myStatement
@@ -185,7 +196,7 @@ class SQLiteAndroidDatabase
                             queryResult = new JSONObject();
                             queryResult.put("rowsAffected", rowsAffected);
                         }
-                    }
+                    // }
 
                     if (needRawQuery) { // for pre-honeycomb behavior
                         rowsAffectedCompat = countRowsAffectedCompat(queryType, query, json_params, mydb);
@@ -466,15 +477,23 @@ class SQLiteAndroidDatabase
                         key = cur.getColumnName(i);
 
                         if (isPostHoneycomb) {
-
                             // Use try & catch just in case android.os.Build.VERSION.SDK_INT >= 11 is lying:
                             try {
                                 bindPostHoneycomb(row, key, cur, i);
                             } catch (Exception ex) {
-                                bindPreHoneycomb(row, key, cur, i);
+                                // bindPreHoneycomb(row, key, cur, i);
+                                Log.v("SQLiteAndroidDatabase.executeSqlStatementQuery",
+                                    "INTERNAL PLUGIN ERROR: could not bindPostHoneycomb: " + ex.getMessage());
+                                throw(ex);
                             }
                         } else {
-                            bindPreHoneycomb(row, key, cur, i);
+                            // NOT EXPECTED:
+                            // bindPreHoneycomb(row, key, cur, i);
+                            Log.v("SQLiteAndroidDatabase.executeSqlStatementQuery",
+                                "INTERNAL PLUGIN ERROR: deprecated android.os.Build.VERSION not supported: " + android.os.Build.VERSION.SDK_INT);
+                            throw new RuntimeException(
+                                "INTERNAL PLUGIN ERROR: deprecated android.os.Build.VERSION not supported: " +
+                                android.os.Build.VERSION.SDK_INT);
                         }
                     }
 
@@ -520,6 +539,7 @@ class SQLiteAndroidDatabase
         }
     }
 
+    /* ** NO LONGER SUPPORTED:
     private void bindPreHoneycomb(JSONObject row, String key, Cursor cursor, int i) throws JSONException {
         // Since cursor.getType() is not available pre-honeycomb, this is
         // a workaround so we don't have to bind everything as a string
@@ -538,6 +558,7 @@ class SQLiteAndroidDatabase
             row.put(key, cursor.getString(i));
         }
     }
+    // */
 
     static QueryType getQueryType(String query) {
         Matcher matcher = FIRST_WORD.matcher(query);